requirements.txt
discord.py==1.7.3
python-dotenv==0.19.0
asyncpraw==7.4.0